+++ src/Network/Network.Test/ScenarioTests/PrivateLinkServiceTests.ps1
@@ -118,3 +118,88 @@ function Test-PrivateLinkServiceCRUD
     }
 } 
 
+<#
+.SYNOPSIS
+Test operation for PrivateEndpointConnection.
+#>
+function Test-PrivateEndpointConnectionCRUD
+{
+    # Setup
+    $rgname = Get-ResourceGroupName;
+    $rname = Get-ResourceName;
+    $location = Get-ProviderLocation "Microsoft.Network/privateLinkServices" "westus2";
+    # Dependency parameters
+    $IpConfigurationName = "IpConfigurationName";
+    $vnetName = Get-ResourceName;
+    $ilbFrontName = "LB-Frontend";
+    $ilbBackendName = "LB-Backend";
+    $ilbName = Get-ResourceName;
+
+    $serverName = Get-ResourceName
+    $serverLogin = "testusername"
+    <#[SuppressMessage("Microsoft.Security", "CS002:SecretInNextLine", Justification="Test passwords only valid for the duration of the test")]#>
+    $serverPassword = "t357ingP@s5w0rd!Sec"
+    $credentials = new-object System.Management.Automation.PSCredential($serverLogin, ($serverPassword | ConvertTo-SecureString -asPlainText -Force))
+    $databaseName = "mySampleDatabase"
+    $peName = "mype"
+    $subId = getSubscription
+
+    try
+    {
+        $resourceGroup = New-AzResourceGroup -Name $rgname -Location $location;
+
+        if ((Get-NetworkTestMode) -ne 'Playback')
+        {
+            # Create Sql Storage Account
+            $server = New-AzSqlServer -ResourceGroupName $rgname `
+                -ServerName $serverName `
+                -Location $location `
+                -SqlAdministratorCredentials $credentials
+
+            $database = New-AzSqlDatabase  -ResourceGroupName $rgname `
+                -ServerName $serverName `
+                -DatabaseName $databaseName `
+                -RequestedServiceObjectiveName "Basic" `
+                -Edition "Basic"
+
+            $sqlResourceId = $server.ResourceId
+        }
+        else
+        {
+            $sqlResourceId = "/subscriptions/$subId/resourceGroups/$rgname/providers/Microsoft.Sql/servers/$serverName"
+        }
+
+        # Create Private Endpoint
+        $peSubnet = New-AzVirtualNetworkSubnetConfig -Name peSubnet -AddressPrefix "11.0.1.0/24" -PrivateEndpointNetworkPolicies "Disabled"
+        $vnetPE = New-AzVirtualNetwork -Name "vnetPE" -ResourceGroupName $rgname -Location $location -AddressPrefix "11.0.0.0/16" -Subnet $peSubnet
+
+        $plsConnection= New-AzPrivateLinkServiceConnection -Name plsConnection -PrivateLinkServiceId  $sqlResourceId -GroupId 'sqlServer'
+        $privateEndpoint = New-AzPrivateEndpoint -ResourceGroupName $rgname -Name $peName -Location $location -Subnet $vnetPE.subnets[0] -PrivateLinkServiceConnection $plsConnection -ByManualRequest
+
+        # Get Private Endpoint Connection
+        $pecGet = Get-AzPrivateEndpointConnection -PrivateLinkResourceId $sqlResourceId
+        Assert-NotNull $pecGet;
+        Assert-AreEqual "Pending" $pecGet.PrivateLinkServiceConnectionState.Status
+
+        # Approve Private Endpoint Connection
+        $pecApprove = Approve-AzPrivateEndpointConnection -ResourceId $pecGet.Id
+        Assert-NotNull $pecApprove;
+        Assert-AreEqual "Approved" $pecApprove.PrivateLinkServiceConnectionState.Status
+
+        Start-TestSleep 30000
+
+        # Remove Private Endpoint Connection
+        $pecRemove = Remove-AzPrivateEndpointConnection -ResourceId $pecGet.Id -PassThru -Force
+        Assert-AreEqual true $pecRemove
+
+        # Get Private Endpoint Connection again
+        $pecGet2 = Get-AzPrivateEndpointConnection -PrivateLinkResourceId $sqlResourceId
+        Assert-Null($pecGet2)
+
+    }
+    finally
+    {
+        # Cleanup
+        Clean-ResourceGroup $rgname;
+    }
+}